Change logstash user to logstash under debian/ubuntu
In Debian and Ubuntu the logstash user and group should be logstash, as the programm runs as that user. Without my change the files will belong to root. 
I found out about this when logstash refused to start because it could not read the pattern file I gave it. It belonged to root with no read permissions for the process.